pyproject.toml
[tool.poetry]
name = "phonology2vec"
version = "0.1.0"
description = "Tools for analyzing and visualizing the phonologies of different languages using corpora written in IPA."
authors = ["Ian Scott Knight <isk@alumni.stanford.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.5"
jupyter = "^1.0.0"
pandas = "^1.1.2"
python-dotenv = "^0.14.0"
scipy = "^1.5.2"
scikit-learn = "^0.23.2"
matplotlib = "^3.3.2"
seaborn = "^0.11.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
